+++ portal-impl/src/com/liferay/portal/service/persistence/AccountPersistenceImpl.java
@@ -238,7 +238,9 @@ public class AccountPersistenceImpl extends BasePersistenceImpl<Account>
 						account.getPrimaryKeyObj());
 			}
 
-			session.delete(account);
+			if (account != null) {
+				session.delete(account);
+			}
 		}
 		catch (Exception e) {
 			throw processException(e);
@@ -247,7 +249,9 @@ public class AccountPersistenceImpl extends BasePersistenceImpl<Account>
 			closeSession(session);
 		}
 
-		clearCache(account);
+		if (account != null) {
+			clearCache(account);
+		}
 
 		return account;
 	}
